Implemented DataLoaderService which automatically calls all available DataLoaders Currently only moved the contents of DataLoader to DefaultDataLoader, it is still one monolith
