+++ backend/src/main/kotlin/outcobra/server/data/DataLoadOrder.kt
@@ -1,0 +1,6 @@
+package outcobra.server.data
+
+object DataLoadOrder {
+    const val DEFAULT_DATALOADER = 1
+    const val TASK_DATALOADER = 2
+}
