+++ backend/src/main/kotlin/outcobra/server/data/DataLoaderService.kt
@@ -1,0 +1,37 @@
+package outcobra.server.data
+
+import org.slf4j.LoggerFactory
+import org.springframework.boot.ApplicationArguments
+import org.springframework.boot.ApplicationRunner
+import org.springframework.stereotype.Component
+import javax.inject.Inject
+
+@Component
+open class DataLoaderService
+@Inject constructor(val dataLoaders: List<DataLoader>) : ApplicationRunner {
+    companion object {
+
+        private val LOGGER = LoggerFactory.getLogger(DataLoaderService::class.java)
+    }
+
+    override fun run(args: ApplicationArguments?) {
+        LOGGER.info("Loading data")
+
+        dataLoaders.forEach { loader ->
+            val loaderName = loader.javaClass.simpleName
+            try {
+                if (loader.shouldLoad()) {
+                    LOGGER.debug("Started loading $loaderName")
+                    loader.load()
+                    LOGGER.debug("Finished loading $loaderName")
+                } else {
+                    LOGGER.info("Not loading $loaderName because shouldLoad() returned false")
+                }
+            } catch (e: Exception) {
+                LOGGER.error("Could not load ${loader.javaClass.simpleName}")
+            }
+        }
+
+        LOGGER.info("Finished loading data")
+    }
+}
