+++ backend/src/main/kotlin/outcobra/server/model/mapper/ColorMapper.kt
@@ -12,6 +12,4 @@ import outcobra.server.model.interfaces.Mapper
 open class ColorMapper : Mapper<Color, ColorDto> {
     override fun fromDto(from: ColorDto): Color? = Color.getByHex(from.hex)
     override fun toDto(from: Color): ColorDto = ColorDto(from.name, from.hex, 0)
-
-
 }
